+++ src/_doc/examples.rs
@@ -14,8 +14,6 @@
 
 #[path = "../../examples/code/enumint.rs"]
 pub mod enumint;
-#[path = "../../examples/code/enumset.rs"]
-pub mod enumset;
 
 #[path = "../../examples/data/id_seq.rs"]
 pub mod id_seq;
@@ -39,13 +37,6 @@ crate::structural_mods! { // _pub_mods
         #[doc(inline)]
         #[doc = crate::_tags!(example num niche)]
         pub use super::enumint::*;
-
-        #[doc(inline)]
-        #[doc = crate::_tags!(example code member)]
-        pub use super::enumset::EnumExample;
-        #[doc(inline)]
-        #[doc = crate::_tags!(example code set)]
-        pub use super::enumset::EnumSetExample;
 
         #[doc(inline)]
         #[doc = crate::_tags!(example uid construction)]