+++ src/data/codec/bin/bit/field/mod.rs
@@ -1,18 +1,18 @@
 // devela/src/data/codec/bin/bit/field/mod.rs
 
 #[cfg(any(test, doctest))]
-mod tests;
+mod _tests;
 
 #[cfg(any(test, feature = "_docs_examples"))]
 mod examples; // BitfieldExample
 
-mod definition; // bitfield!
+mod define; // bitfield!
 
 crate::structural_mods! { // _mods
     _mods {
         #[doc(inline)]
         pub use super::{
-            definition::bitfield,
+            define::bitfield,
         };
         #[cfg(any(test, feature = "_docs_examples"))]
         pub use super::examples::*;