+++ src/data/codec/bin/bit/set/mod.rs
@@ -1,15 +1,15 @@
 // devela/src/data/codec/bin/bit/set/mod.rs
 
 #[cfg(any(test, doctest))]
-mod tests;
+mod _tests;
 
-mod definition; // set!
+mod define; // set!
 
 crate::structural_mods! { // _mods
     _mods {
         #[doc(inline)]
         pub use super::{
-            definition::set,
+            define::set,
         };
     }
 }