+++ docs/changelog.md
@@ -29,6 +29,7 @@
 - clarify public module and hidden scope features.
 
 ## examples
+- remove examples: `enumset`.
 - update example: `term_linux`:
   - showcase signal handling.
 - update examples: `web_api`, `web_workers`: